Update schema and endpoints
Previously, we had planned to update each field as the user populated
it, using an unvalidated staging table, so that they could pause
completing their RSVP and resume on another device or at another time
before submitting.

However, we determined it would add too much additional code complexity
to maintain the integrity of submitted data in the face of race
conditions given multiple concurrent instances of the form being
updated.

As such, the server-side database will only be updated once the form has
been submitted.

It is worth noting that the constraint on the RSVP table barring brunch
RSVPs from non-invited guests is not safe were there to be updates to
the invitation status of records in the guest table while the trigger is
being executed.